Moved `getVariableReferences` from Operation to Document
