+++ layers/Engine/packages/graphql-parser/src/Spec/Parser/Ast/Document.php
@@ -7,6 +7,9 @@ namespace PoP\GraphQLParser\Spec\Parser\Ast;
 use PoP\GraphQLParser\Error\GraphQLErrorMessageProviderInterface;
 use PoP\GraphQLParser\Exception\Parser\InvalidRequestException;
 use PoP\GraphQLParser\Facades\Error\GraphQLErrorMessageProviderFacade;
+use PoP\GraphQLParser\Spec\Parser\Ast\ArgumentValue\InputList;
+use PoP\GraphQLParser\Spec\Parser\Ast\ArgumentValue\InputObject;
+use PoP\GraphQLParser\Spec\Parser\Ast\ArgumentValue\VariableReference;
 use PoP\GraphQLParser\Spec\Parser\Location;
 use PoP\Root\Services\StandaloneServiceTrait;
 
@@ -309,6 +312,134 @@ class Document implements DocumentInterface
                 );
             }
         }
+    }
+
+    /**
+     * Gather all the VariableReference within the Operation.
+     *
+     * @param Fragment[] $fragments
+     * @return VariableReference[]
+     */
+    protected function getVariableReferences(array $fragments): array
+    {
+        return array_merge(
+            $this->getVariableReferencesInFieldsOrFragments($this->fieldsOrFragmentBonds, $fragments),
+            $this->getVariableReferencesInDirectives($this->directives)
+        );
+    }
+
+    /**
+     * @param FieldInterface[]|FragmentBondInterface[] $fieldsOrFragmentBonds
+     * @param Fragment[] $fragments
+     * @return VariableReference[]
+     */
+    protected function getVariableReferencesInFieldsOrFragments(array $fieldsOrFragmentBonds, array $fragments): array
+    {
+        $variableReferences = [];
+        foreach ($fieldsOrFragmentBonds as $fieldOrFragmentBond) {
+            if ($fieldOrFragmentBond instanceof FragmentReference) {
+                /** @var FragmentReference */
+                $fragmentReference = $fieldOrFragmentBond;
+                $fragment = $this->getFragment($fragments, $fragmentReference->getName());
+                if ($fragment === null) {
+                    continue;
+                }
+                $variableReferences = array_merge(
+                    $variableReferences,
+                    $this->getVariableReferencesInFieldsOrFragments($fragment->getFieldsOrFragmentBonds(), $fragments)
+                );
+                continue;
+            }
+            if ($fieldOrFragmentBond instanceof InlineFragment) {
+                /** @var InlineFragment */
+                $inlineFragment = $fieldOrFragmentBond;
+                $variableReferences = array_merge(
+                    $variableReferences,
+                    $this->getVariableReferencesInFieldsOrFragments($inlineFragment->getFieldsOrFragmentBonds(), $fragments)
+                );
+                continue;
+            }
+            /** @var FieldInterface */
+            $field = $fieldOrFragmentBond;
+            $variableReferences = array_merge(
+                $variableReferences,
+                $this->getVariableReferencesInArguments($field->getArguments()),
+                $this->getVariableReferencesInDirectives($field->getDirectives())
+            );
+            if ($field instanceof RelationalField) {
+                /** @var RelationalField */
+                $relationalField = $field;
+                $variableReferences = array_merge(
+                    $variableReferences,
+                    $this->getVariableReferencesInFieldsOrFragments($relationalField->getFieldsOrFragmentBonds(), $fragments)
+                );
+                continue;
+            }
+        }
+        return $variableReferences;
+    }
+
+    /**
+     * @param Directive[] $directives
+     * @return VariableReference[]
+     */
+    protected function getVariableReferencesInDirectives(array $directives): array
+    {
+        $variableReferences = [];
+        foreach ($directives as $directive) {
+            $variableReferences = array_merge(
+                $variableReferences,
+                $this->getVariableReferencesInArguments($directive->getArguments())
+            );
+        }
+        return $variableReferences;
+    }
+
+    /**
+     * @param Argument[] $arguments
+     * @return VariableReference[]
+     */
+    protected function getVariableReferencesInArguments(array $arguments): array
+    {
+        $variableReferences = [];
+        foreach ($arguments as $argument) {
+            $variableReferences = array_merge(
+                $variableReferences,
+                $this->getVariableReferencesInArgumentValue($argument->getValue())
+            );
+        }
+        return $variableReferences;
+    }
+
+    /**
+     * @return VariableReference[]
+     */
+    protected function getVariableReferencesInArgumentValue(WithValueInterface $argumentValue): array
+    {
+        if ($argumentValue instanceof VariableReference) {
+            return [$argumentValue];
+        }
+        if (!($argumentValue instanceof InputObject || $argumentValue instanceof InputList)) {
+            return [];
+        }
+        // Get references within InputObjects and Lists
+        $variableReferences = [];
+        $listValues = (array)$argumentValue->getAstValue();
+        foreach ($listValues as $listValue) {
+            if (!($listValue instanceof VariableReference || $listValue instanceof WithValueInterface)) {
+                continue;
+            }
+            if ($listValue instanceof VariableReference) {
+                $variableReferences[] = $listValue;
+                continue;
+            }
+            /** @var WithValueInterface $listValue */
+            $variableReferences = array_merge(
+                $variableReferences,
+                $this->getVariableReferencesInArgumentValue($listValue)
+            );
+        }
+        return $variableReferences;
     }
 
     /**
