+++ layers/Engine/packages/graphql-parser/src/Spec/Parser/Ast/OperationInterface.php
@@ -16,10 +16,4 @@ interface OperationInterface extends LocatableInterface, WithDirectivesInterface
      * @return Variable[]
      */
     public function getVariables(): array;
-
-    /**
-     * @param Fragment[] $fragments
-     * @return VariableReference[]
-     */
-    public function getVariableReferences(array $fragments): array;
 }
